Revert "Fixing to work with the most current artwork, started to integrate the multi-huds feature, added installer from ACSS, added a whole lot of R8 features."
This reverts commit ec359ae190476fb47b67dc9c68d9effa91239c1b.

+++ src/zlib/color.cpp
@@ -8,7 +8,7 @@
 
 float HueComponent(float hue)
 {
-    hue = (float)fabs(mod(hue + 3, 6) - 3);
+    hue = fabs(mod(hue + 3, 6) - 3);
 
     if (hue > 2) {
         return 0;
